+++ system/platform-core/src/main/java/org/platformlambda/core/annotations/PreLoad.java
@@ -43,5 +43,4 @@ public @interface PreLoad {
     int instances() default 1;
     String envInstances() default "";
     boolean isPrivate() default true;
-
 }